requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
boto3==1.37.14
botocore==1.37.14
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.13
django-celery-results==2.5.1
django-environ==0.12.0
django-redis==5.4.0
django-storages==1.14.5
jmespath==1.0.1
kombu==5.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==5.2.1
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
gunicorn
Pillow==11.1.0
django-modeltranslation==0.19.12
